Fixed bug with svgr options for react native + adding example color elements + opacity
